# Conflicts:
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/delhivery/axle/api/response/FastagRechargeResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class FastagRechargeResponse(
     @SerializedName("fastag_id") val fastagId: String?,
-    @SerializedName("recharge_amount") val rechargeAmount: Int?,
+    @SerializedName("recharge_amount") val rechargeAmount: Double?,
     @SerializedName("updated_wallet_balance") val updatedWalletBalance: Double?,
     @SerializedName("fastag_balance") val fastagBalance: String?,
     @SerializedName("status") val status: String?
